chore: editorial pass on the package README
Summary:

This adds some clarity to the README as to the purpose of this library, why
one might want to use it, and how one would use it. I also eliminated all
mention of "foreground" and "background" running modes in favour of a
more general description of the Adaptor lifecycle and the CLI subcommands.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>